heavy file cleanup
