+++ src/components/display-panel/attackbar.tsx
