+++ src/components/display-panel/healthbar.tsx
