--- src/components/footer.tsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-const ScWrapper = styled.footer`
-`;
-
-function Footer() {
-  return (
-    <ScWrapper>
-    </ScWrapper>
-  );
-}
-
-export default Footer;
