--- src/components/header.tsx
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-const ScWrapper = styled.header`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-`;
-
-function Header() {
-  return (
-    <ScWrapper>
-    </ScWrapper>
-  );
-}
-
-export default Header;
